Move tests to independent file, and add detailed tests on NaN propagation
