+++ tests/test_raster/test_raster.py
@@ -10,7 +10,6 @@ import tempfile
 import warnings
 from io import StringIO
 from tempfile import TemporaryFile
-from typing import Literal
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -19,15 +18,11 @@ import rasterio as rio
 import xarray as xr
 from pylint.lint import Run
 from pylint.reporters.text import TextReporter
-from scipy.ndimage import distance_transform_edt
 
 import geoutils as gu
-import geoutils.projtools as pt
 from geoutils import examples
 from geoutils._typing import MArrayNum, NDArrayNum
 from geoutils.misc import resampling_method_from_str
-from geoutils.projtools import reproject_to_latlon
-from geoutils.raster.interpolate import method_to_order
 from geoutils.raster.raster import _default_nodata, _default_rio_attrs
 
 DO_PLOT = False
@@ -1982,519 +1977,6 @@ class TestRaster:
             intersection = r.intersection(r_nonoverlap)
             assert intersection == (0.0, 0.0, 0.0, 0.0)
 
-    @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path, landsat_rgb_path])  # type: ignore
-    def test_ij2xy(self, example: str) -> None:
-        """Test the outputs of ij2xy and that the function is reversible with xy2ij."""
-
-        # Open raster
-        rst = gu.Raster(example)
-        xmin, ymin, xmax, ymax = rst.bounds
-
-        # Check ij2xy manually for the four corners and center
-        # With "force_offset", no considerations of pixel interpolation
-
-        # With offset="center", should be pixel center
-        xmin_center = xmin + rst.res[0] / 2
-        ymin_center = ymin + rst.res[1] / 2
-        xmax_center = xmax - rst.res[0] / 2
-        ymax_center = ymax - rst.res[1] / 2
-        assert rst.ij2xy([0], [0], force_offset="center") == ([xmin_center], [ymax_center])
-        assert rst.ij2xy([rst.shape[0] - 1], [0], force_offset="center") == ([xmin_center], [ymin_center])
-        assert rst.ij2xy([0], [rst.shape[1] - 1], force_offset="center") == ([xmax_center], [ymax_center])
-        assert rst.ij2xy([rst.shape[0] - 1], [rst.shape[1] - 1], force_offset="center") == (
-            [xmax_center],
-            [ymin_center],
-        )
-
-        # Same checks for offset="ll", "ul", "ur", "lr
-        lims_ll = [xmin, ymin, xmax - rst.res[0], ymax - rst.res[1]]
-        lims_ul = [xmin, ymin + rst.res[1], xmax - rst.res[0], ymax]
-        lims_lr = [xmin + rst.res[0], ymin, xmax, ymax - rst.res[1]]
-        lims_ur = [xmin + rst.res[0], ymin + rst.res[1], xmax, ymax]
-        offsets = ["ll", "ul", "lr", "ur"]
-        list_lims = [lims_ll, lims_ul, lims_lr, lims_ur]
-        for i in range(len(list_lims)):
-            offset = offsets[i]
-            lim = list_lims[i]
-            assert rst.ij2xy([0], [0], force_offset=offset) == ([lim[0]], [lim[3]])
-            assert rst.ij2xy([rst.shape[0] - 1], [0], force_offset=offset) == ([lim[0]], [lim[1]])
-            assert rst.ij2xy([0], [rst.shape[1] - 1], force_offset=offset) == ([lim[2]], [lim[3]])
-            assert rst.ij2xy([rst.shape[0] - 1], [rst.shape[1] - 1], force_offset=offset) == ([lim[2]], [lim[1]])
-
-        # Check that default coordinate is upper-left
-
-        # With shift from pixel interpretation, coordinates will be half a pixel back for "Point"
-        if rst.area_or_point is not None and rst.area_or_point == "Point" and gu.config["shift_area_or_point"]:
-            # Shift is backward in X, forward in Y
-            lims_ul[0] = lims_ul[0] - 0.5 * rst.res[0]
-            lims_ul[1] = lims_ul[1] + 0.5 * rst.res[1]
-            lims_ul[2] = lims_ul[2] - 0.5 * rst.res[0]
-            lims_ul[3] = lims_ul[3] + 0.5 * rst.res[1]
-
-        assert rst.ij2xy([0], [0]) == ([lims_ul[0]], [lims_ul[3]])
-        assert rst.ij2xy([rst.shape[0] - 1], [0]) == ([lims_ul[0]], [lims_ul[1]])
-        assert rst.ij2xy([0], [rst.shape[1] - 1]) == ([lims_ul[2]], [lims_ul[3]])
-        assert rst.ij2xy([rst.shape[0] - 1], [rst.shape[1] - 1]) == ([lims_ul[2]], [lims_ul[1]])
-
-    @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path, landsat_rgb_path])  # type: ignore
-    def test_xy2ij_ij2xy_reversible(self, example: str):
-
-        # Open raster
-        rst = gu.Raster(example)
-        xmin, ymin, xmax, ymax = rst.bounds
-
-        # We generate random points within the boundaries of the image
-        rng = np.random.default_rng(42)
-        xrand = rng.integers(low=0, high=rst.width, size=(10,)) * list(rst.transform)[0] + xmin
-        yrand = ymax + rng.integers(low=0, high=rst.height, size=(10,)) * list(rst.transform)[4]
-
-        # Test reversibility for any point or area interpretation
-        i, j = rst.xy2ij(xrand, yrand)
-        xnew, ynew = rst.ij2xy(i, j)
-        assert all(xnew == xrand)
-        assert all(ynew == yrand)
-
-        # TODO: clarify this weird behaviour of rasterio.index with floats?
-        # r.ds.index(x, y)
-        # Out[33]: (75, 301)
-        # r.ds.index(x, y, op=np.float32)
-        # Out[34]: (75.0, 302.0)
-
-    def test_xy2ij(self) -> None:
-        """Test xy2ij with shift_area_or_point argument, and compare to interp_points function for consistency."""
-
-        # First, we try on a Raster with a Point interpretation in its "AREA_OR_POINT" metadata: values interpolated
-        # at the center of pixel
-        r = gu.Raster(self.landsat_b4_path)
-        assert r.area_or_point == "Point"
-        xmin, ymin, xmax, ymax = r.bounds
-
-        # We generate random points within the boundaries of the image
-        rng = np.random.default_rng(42)
-        xrand = rng.integers(low=0, high=r.width, size=(10,)) * list(r.transform)[0] + xmin
-        yrand = ymax + rng.integers(low=0, high=r.height, size=(10,)) * list(r.transform)[4]
-
-        # Get decimal indexes based on "Point", should refer to the corner still (shift False by default)
-        i, j = r.xy2ij(xrand, yrand, shift_area_or_point=False)
-        assert np.all(i % 1 == 0)
-        assert np.all(j % 1 == 0)
-
-        # Those should all be .5 because values refer to the center and are shifted
-        i, j = r.xy2ij(xrand, yrand, shift_area_or_point=True)
-        assert np.all(i % 1 == 0.5)
-        assert np.all(j % 1 == 0.5)
-
-        # Force "Area", should refer to corner
-        r.set_area_or_point("Area", shift_area_or_point=False)
-        i, j = r.xy2ij(xrand, yrand, shift_area_or_point=True)
-        assert np.all(i % 1 == 0)
-        assert np.all(j % 1 == 0)
-
-        # Now, we calculate the mean of values in each 2x2 slices of the data, and compare with interpolation at order 1
-        list_z_ind = []
-        img = r.data
-        for k in range(len(xrand)):
-            # 2x2 slices
-            z_ind = np.mean(
-                img[
-                    slice(int(np.floor(i[k])), int(np.ceil(i[k])) + 1),
-                    slice(int(np.floor(j[k])), int(np.ceil(j[k])) + 1),
-                ]
-            )
-            list_z_ind.append(z_ind)
-
-        # First order interpolation
-        rpts = r.interp_points((xrand, yrand), method="linear")
-        # The values interpolated should be equal
-        assert np.array_equal(np.array(list_z_ind, dtype=np.float32), rpts, equal_nan=True)
-
-        # Test there is no failure with random coordinates (edge effects, etc)
-        xrand = rng.uniform(low=xmin, high=xmax, size=(1000,))
-        yrand = rng.uniform(low=ymin, high=ymax, size=(1000,))
-        r.interp_points((xrand, yrand))
-
-        # Second, test after a crop: the Raster now has an Area interpretation, those should fall right on the integer
-        # pixel indexes
-        r2 = gu.Raster(self.landsat_b4_crop_path)
-        r.crop(r2)
-        assert r.area_or_point == "Area"
-
-        xmin, ymin, xmax, ymax = r.bounds
-
-        # We can test with several method for the exact indexes: interp, and simple read should
-        # give back the same values that fall right on the coordinates
-        xrand = rng.integers(low=0, high=r.width, size=(10,)) * list(r.transform)[0] + xmin
-        yrand = ymax + rng.integers(low=0, high=r.height, size=(10,)) * list(r.transform)[4]
-        # By default, i and j are returned as integers
-        i, j = r.xy2ij(xrand, yrand, op=np.float32)
-        list_z_ind = []
-        img = r.data
-        for k in range(len(xrand)):
-            # We directly sample the values
-            z_ind = img[int(i[k]), int(j[k])]
-            list_z_ind.append(z_ind)
-
-        rpts = r.interp_points((xrand, yrand), method="linear")
-
-        assert np.array_equal(np.array(list_z_ind, dtype=np.float32), rpts, equal_nan=True)
-
-    @pytest.mark.parametrize("tag_aop", [None, "Area", "Point"])  # type: ignore
-    @pytest.mark.parametrize("shift_aop", [True, False])  # type: ignore
-    def test_interp_points__synthetic(self, tag_aop: str | None, shift_aop: bool) -> None:
-        """Test interp_points function with synthetic data."""
-
-        # We flip the array up/down to facilitate index comparison of Y axis
-        arr = np.flipud(np.array([1, 2, 3, 4, 5, 6, 7, 8, 9]).reshape((3, 3)))
-        transform = rio.transform.from_bounds(0, 0, 3, 3, 3, 3)
-        raster = gu.Raster.from_array(data=arr, transform=transform, crs=None, nodata=-9999)
-
-        # Define the AREA_OR_POINT attribute without re-transforming
-        raster.set_area_or_point(tag_aop, shift_area_or_point=False)
-
-        # Check interpolation falls right on values for points (1, 1), (1, 2) etc...
-        index_x = [0, 1, 2, 0, 1, 2, 0, 1, 2]
-        index_y = [0, 0, 0, 1, 1, 1, 2, 2, 2]
-
-        # The actual X/Y coords will be offset by one because Y axis is inverted and pixel coords is upper-left corner
-        points_x, points_y = raster.ij2xy(i=index_x, j=index_y, shift_area_or_point=shift_aop)
-
-        # The following 4 methods should yield the same result because:
-        # Nearest = Linear interpolation at the location of a data point
-        # Regular grid = Equal grid interpolation at the location of a data point
-
-        raster_points = raster.interp_points((points_x, points_y), method="nearest", shift_area_or_point=shift_aop)
-        raster_points_lin = raster.interp_points((points_x, points_y), method="linear", shift_area_or_point=shift_aop)
-        raster_points_interpn = raster.interp_points(
-            (points_x, points_y), method="nearest", force_scipy_function="interpn", shift_area_or_point=shift_aop
-        )
-        raster_points_interpn_lin = raster.interp_points(
-            (points_x, points_y), method="linear", force_scipy_function="interpn", shift_area_or_point=shift_aop
-        )
-
-        assert np.array_equal(raster_points, raster_points_lin)
-        assert np.array_equal(raster_points, raster_points_interpn)
-        assert np.array_equal(raster_points, raster_points_interpn_lin)
-
-        for i in range(3):
-            for j in range(3):
-                ind = 3 * i + j
-                assert raster_points[ind] == arr[index_x[ind], index_y[ind]]
-
-        # Check bilinear interpolation values inside the grid (same here, offset by 1 between X and Y)
-        index_x_in = [0.5, 0.5, 1.5, 1.5]
-        index_y_in = [0.5, 1.5, 0.5, 1.5]
-
-        points_x_in, points_y_in = raster.ij2xy(i=index_x_in, j=index_y_in, shift_area_or_point=shift_aop)
-
-        # Here again compare methods
-        raster_points_in = raster.interp_points(
-            (points_x_in, points_y_in), method="linear", shift_area_or_point=shift_aop
-        )
-        raster_points_in_interpn = raster.interp_points(
-            (points_x_in, points_y_in), method="linear", force_scipy_function="interpn", shift_area_or_point=shift_aop
-        )
-
-        assert np.array_equal(raster_points_in, raster_points_in_interpn)
-
-        for i in range(len(points_x_in)):
-
-            xlow = int(index_x_in[i] - 0.5)
-            xupp = int(index_x_in[i] + 0.5)
-            ylow = int(index_y_in[i] - 0.5)
-            yupp = int(index_y_in[i] + 0.5)
-
-            # Check the bilinear interpolation matches the mean value of those 4 points (equivalent as its the middle)
-            assert raster_points_in[i] == np.mean([arr[xlow, ylow], arr[xupp, ylow], arr[xupp, yupp], arr[xlow, yupp]])
-
-        # Check bilinear extrapolation for points at 1 spacing outside from the input grid
-        points_out = (
-            [(-1, i) for i in np.arange(1, 4)]
-            + [(i, -1) for i in np.arange(1, 4)]
-            + [(4, i) for i in np.arange(1, 4)]
-            + [(i, 4) for i in np.arange(4, 1)]
-        )
-        points_out_xy = list(zip(*points_out))
-        raster_points_out = raster.interp_points(points_out_xy)
-        assert all(~np.isfinite(raster_points_out))
-
-        # To use cubic or quintic, we need a larger grid (minimum 6x6, but let's aim bigger with 50x50)
-        arr = np.flipud(np.arange(1, 2501).reshape((50, 50)))
-        transform = rio.transform.from_bounds(0, 0, 50, 50, 50, 50)
-        raster = gu.Raster.from_array(data=arr, transform=transform, crs=None, nodata=-9999)
-        raster.set_area_or_point(tag_aop, shift_area_or_point=False)
-
-        # For this, get random points
-        rng = np.random.default_rng(42)
-        index_x_in_rand = rng.integers(low=8, high=42, size=(10,)) + rng.normal(scale=0.3)
-        index_y_in_rand = rng.integers(low=8, high=42, size=(10,)) + rng.normal(scale=0.3)
-        points_x_rand, points_y_rand = raster.ij2xy(i=index_x_in_rand, j=index_y_in_rand, shift_area_or_point=shift_aop)
-
-        for method in ["nearest", "linear"]:
-            raster_points_mapcoords = raster.interp_points(
-                (points_x_rand, points_y_rand),
-                method=method,
-                force_scipy_function="map_coordinates",
-                shift_area_or_point=shift_aop,
-            )
-            raster_points_interpn = raster.interp_points(
-                (points_x_rand, points_y_rand),
-                method=method,
-                force_scipy_function="interpn",
-                shift_area_or_point=shift_aop,
-            )
-
-            # Not exactly equal in floating point precision since changes in Scipy 1.13.0,
-            # see https://github.com/GlacioHack/geoutils/issues/533
-            assert np.allclose(raster_points_mapcoords, raster_points_interpn, atol=0.01)
-
-        # Check that, outside the edge, the interpolation fails and returns a NaN
-        index_x_edge_rand = [-0.5, -0.5, -0.5, 25, 25, 49.5, 49.5, 49.5]
-        index_y_edge_rand = [-0.5, 25, 49.5, -0.5, 49.5, -0.5, 25, 49.5]
-
-        points_x_rand, points_y_rand = raster.ij2xy(
-            i=index_x_edge_rand, j=index_y_edge_rand, shift_area_or_point=shift_aop
-        )
-
-        # Nearest doesn't apply, just linear and above
-        for method in ["cubic", "quintic"]:
-            raster_points_mapcoords_edge = raster.interp_points(
-                (points_x_rand, points_y_rand),
-                method=method,
-                force_scipy_function="map_coordinates",
-                shift_area_or_point=shift_aop,
-            )
-            raster_points_interpn_edge = raster.interp_points(
-                (points_x_rand, points_y_rand),
-                method=method,
-                force_scipy_function="interpn",
-                shift_area_or_point=shift_aop,
-            )
-
-            assert all(~np.isfinite(raster_points_mapcoords_edge))
-            assert all(~np.isfinite(raster_points_interpn_edge))
-
-    @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path])  # type: ignore
-    @pytest.mark.parametrize(
-        "method", ["nearest", "linear", "cubic", "quintic", "slinear", "pchip", "splinef2d"]
-    )  # type: ignore
-    def test_interp_points__real(
-        self, example: str, method: Literal["nearest", "linear", "cubic", "quintic", "slinear", "pchip", "splinef2d"]
-    ) -> None:
-        """Test interp_points for real data."""
-
-        # 1/ Check the accuracy of the interpolation at an exact point, and between methods
-
-        r = gu.Raster(example)
-        r.set_area_or_point("Area", shift_area_or_point=False)
-
-        # Test for an individual point (shape can be tricky at 1 dimension)
-        itest = 100
-        jtest = 100
-        x, y = r.ij2xy(itest, jtest)
-        val = r.interp_points((x, y), method=method, force_scipy_function="map_coordinates")[0]
-        val_img = r.data[itest, jtest]
-        if "nearest" in method or "linear" in method:
-            assert val_img == val
-
-        # Check the result is exactly the same for both methods
-        val2 = r.interp_points((x, y), method=method, force_scipy_function="interpn")[0]
-        assert val2 == pytest.approx(val)
-
-        # Check that interp convert to latlon
-        lat, lon = gu.projtools.reproject_to_latlon([x, y], in_crs=r.crs)
-        val_latlon = r.interp_points((lat, lon), method=method, input_latlon=True)[0]
-        assert val == pytest.approx(val_latlon, abs=0.0001)
-
-        # 2/ Check the propagation of NaNs
-
-        # Convert raster to float
-        r = r.astype(np.float32)
-
-        # Create a NaN at a given pixel (we know the landsat example has no NaNs to begin with)
-        i0, j0 = (10, 10)
-        r[i0, j0] = np.nan
-
-        # All surrounding pixels with distance half the method order rounded up should be NaNs
-        order = method_to_order[method]
-        if method == "linear":
-            order = 1
-        d = int(np.ceil(order / 2))
-        # No NaN propagation for linear
-        indices_nan = [
-            (i0 + i, j0 + j) for i in np.arange(-d, d + 1) for j in np.arange(-d, d + 1) if (np.abs(i) + np.abs(j)) <= d
-        ]
-        i, j = list(zip(*indices_nan))
-        x, y = r.ij2xy(i, j)
-        vals = r.interp_points((x, y), method=method, force_scipy_function="map_coordinates")[0]
-        vals2 = r.interp_points((x, y), method=method, force_scipy_function="interpn")[0]
-
-        assert all(np.isnan(np.atleast_1d(vals))) and all(np.isnan(np.atleast_1d(vals2)))
-
-        # 3/ Check that valid interpolated values at the edge of NaNs are free of errors
-
-        # We compare values interpolated right at the edge of valid values near a NaN between
-        # 1/ Implementation of interp_points (that replaces NaNs by a placeholder value of 0 during interpolation)
-        # 2/ Raster filled with nearest neighbour at NaN coordinates, then running interp_points
-        # If the interpolated value are free of the influence of the placeholder value of 0, the interpolated value
-        # should be exactly the same
-
-        # We get the indexes of valid pixels just at the edge of NaNs
-        indices_edge = [
-            (i0 + i, j0 + j)
-            for i in np.arange(-d - 1, d + 2)
-            for j in np.arange(-d - 1, d + 2)
-            if (np.abs(i) + np.abs(j)) == d + 1
-        ]
-        i, j = list(zip(*indices_edge))
-        x, y = r.ij2xy(i, j)
-        # And get their interpolated value
-        vals = r.interp_points((x, y), method=method, force_scipy_function="map_coordinates")[0]
-        vals2 = r.interp_points((x, y), method=method, force_scipy_function="interpn")[0]
-
-        # Then we fill the NaNs in the raster with the nearest neighbour
-        r_arr = r.get_nanarray()
-        # Elegant solution from: https://stackoverflow.com/questions/5551286/filling-gaps-in-a-numpy-array
-        indices = distance_transform_edt(~np.isfinite(r_arr), return_distances=False, return_indices=True)
-        r_arr = r_arr[tuple(indices)]
-        r.data = r_arr
-
-        # All raster values should be valid now
-        assert np.all(np.isfinite(r_arr))
-
-        # And get the interpolated values
-        vals_near = r.interp_points((x, y), method=method, force_scipy_function="map_coordinates")[0]
-        vals2_near = r.interp_points((x, y), method=method, force_scipy_function="interpn")[0]
-
-        # Both sets of values should be exactly the same, without any NaNs
-        assert np.allclose(vals, vals_near, equal_nan=False)
-        assert np.allclose(vals2, vals2_near, equal_nan=False)
-
-    def test_value_at_coords(self) -> None:
-        """
-        Test that value at coords works as intended
-        """
-
-        # -- Tests 1: check based on indexed values --
-
-        # Open raster
-        r = gu.Raster(self.landsat_b4_crop_path)
-
-        # A pixel center where all neighbouring coordinates are different:
-        # array([[[237, 194, 239],
-        #          [250, 173, 164],
-        #          [255, 192, 128]]]
-        itest0 = 120
-        jtest0 = 451
-        # This is the center of the pixel
-        xtest0 = 496975.0
-        ytest0 = 3099095.0
-
-        # Verify coordinates match indexes
-        x_out, y_out = r.ij2xy(itest0, jtest0, force_offset="center")
-        assert x_out == xtest0
-        assert y_out == ytest0
-
-        # Check that the value at this coordinate is the same as when indexing
-        z_val = r.value_at_coords(xtest0, ytest0)
-        z = r.data.data[itest0, jtest0]
-        assert z == z_val
-
-        # Check that the value is the same the other 4 corners of the pixel
-        assert z == r.value_at_coords(xtest0 + 0.49 * r.res[0], ytest0 - 0.49 * r.res[1])
-        assert z == r.value_at_coords(xtest0 - 0.49 * r.res[0], ytest0 + 0.49 * r.res[1])
-        assert z == r.value_at_coords(xtest0 - 0.49 * r.res[0], ytest0 - 0.49 * r.res[1])
-        assert z == r.value_at_coords(xtest0 + 0.49 * r.res[0], ytest0 + 0.49 * r.res[1])
-
-        # -- Tests 2: check arguments work as intended --
-
-        # 1/ Lat-lon argument check by getting the coordinates of our last test point
-        lat, lon = reproject_to_latlon(points=[xtest0, ytest0], in_crs=r.crs)
-        z_val_2 = r.value_at_coords(lon, lat, latlon=True)
-        assert z_val == z_val_2
-
-        # 2/ Band argument
-        # Get the band indexes for the multi-band Raster
-        r_multi = gu.Raster(self.landsat_rgb_path)
-        itest, jtest = r_multi.xy2ij(xtest0, ytest0)
-        itest = int(itest[0])
-        jtest = int(jtest[0])
-        # Extract the values
-        z_band1 = r_multi.value_at_coords(xtest0, ytest0, band=1)
-        z_band2 = r_multi.value_at_coords(xtest0, ytest0, band=2)
-        z_band3 = r_multi.value_at_coords(xtest0, ytest0, band=3)
-        # Compare to the Raster array slice
-        assert list(r_multi.data[:, itest, jtest]) == [z_band1, z_band2, z_band3]
-
-        # 3/ Masked argument
-        r_multi.data[:, itest, jtest] = np.ma.masked
-        z_not_ma = r_multi.value_at_coords(xtest0, ytest0, band=1)
-        assert not np.ma.is_masked(z_not_ma)
-        z_ma = r_multi.value_at_coords(xtest0, ytest0, band=1, masked=True)
-        assert np.ma.is_masked(z_ma)
-
-        # 4/ Window argument
-        val_window, z_window = r_multi.value_at_coords(
-            xtest0, ytest0, band=1, window=3, masked=True, return_window=True
-        )
-        assert (
-            val_window
-            == np.ma.mean(r_multi.data[0, itest - 1 : itest + 2, jtest - 1 : jtest + 2])
-            == np.ma.mean(z_window)
-        )
-        assert np.array_equal(z_window, r_multi.data[0, itest - 1 : itest + 2, jtest - 1 : jtest + 2])
-
-        # 5/ Reducer function argument
-        val_window2 = r_multi.value_at_coords(
-            xtest0, ytest0, band=1, window=3, masked=True, reducer_function=np.ma.median
-        )
-        assert val_window2 == np.ma.median(r_multi.data[0, itest - 1 : itest + 2, jtest - 1 : jtest + 2])
-
-        # -- Tests 3: check that errors are raised when supposed for non-boolean arguments --
-
-        # Verify that passing a window that is not a whole number fails
-        with pytest.raises(ValueError, match=re.escape("Window must be a whole number.")):
-            r.value_at_coords(xtest0, ytest0, window=3.5)  # type: ignore
-        # Same for an odd number
-        with pytest.raises(ValueError, match=re.escape("Window must be an odd number.")):
-            r.value_at_coords(xtest0, ytest0, window=4)
-        # But a window that is a whole number as a float works
-        r.value_at_coords(xtest0, ytest0, window=3.0)  # type: ignore
-
-        # -- Tests 4: check that passing an array-like object works
-
-        # For simple coordinates
-        x_coords = [xtest0, xtest0 + 100]
-        y_coords = [ytest0, ytest0 - 100]
-        vals = r_multi.value_at_coords(x=x_coords, y=y_coords)
-        val0, win0 = r_multi.value_at_coords(x=x_coords[0], y=y_coords[0], return_window=True)
-        val1, win1 = r_multi.value_at_coords(x=x_coords[1], y=y_coords[1], return_window=True)
-
-        assert len(vals) == len(x_coords)
-        assert vals[0] == val0
-        assert vals[1] == val1
-
-        # With a return window argument
-        vals, windows = r_multi.value_at_coords(x=x_coords, y=y_coords, return_window=True)
-        assert len(windows) == len(x_coords)
-        assert np.array_equal(windows[0], win0, equal_nan=True)
-        assert np.array_equal(windows[1], win1, equal_nan=True)
-
-        # -- Tests 5 -- Check image corners and latlon argument
-
-        # Lower right pixel
-        x, y = [r.bounds.right - r.res[0] / 2, r.bounds.bottom + r.res[1] / 2]
-        lat, lon = pt.reproject_to_latlon([x, y], r.crs)
-        assert r.value_at_coords(x, y) == r.value_at_coords(lon, lat, latlon=True) == r.data[-1, -1]
-
-        # One pixel above
-        x, y = [r.bounds.right - r.res[0] / 2, r.bounds.bottom + 3 * r.res[1] / 2]
-        lat, lon = pt.reproject_to_latlon([x, y], r.crs)
-        assert r.value_at_coords(x, y) == r.value_at_coords(lon, lat, latlon=True) == r.data[-2, -1]
-
-        # One pixel left
-        x, y = [r.bounds.right - 3 * r.res[0] / 2, r.bounds.bottom + r.res[1] / 2]
-        lat, lon = pt.reproject_to_latlon([x, y], r.crs)
-        assert r.value_at_coords(x, y) == r.value_at_coords(lon, lat, latlon=True) == r.data[-1, -2]
 
     @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path])  # type: ignore
     def test_set_nodata(self, example: str) -> None:
@@ -2963,46 +2445,6 @@ class TestRaster:
             temp_dir.cleanup()
         except (NotADirectoryError, PermissionError):
             pass
-
-    @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path, landsat_rgb_path])  # type: ignore
-    def test_coords(self, example: str) -> None:
-
-        img = gu.Raster(self.landsat_b4_path)
-
-        # With lower left argument
-        xx0, yy0 = img.coords(grid=False, force_offset="ll")
-
-        assert len(xx0) == img.width
-        assert len(yy0) == img.height
-
-        assert xx0[0] == pytest.approx(img.bounds.left)
-        assert xx0[-1] == pytest.approx(img.bounds.right - img.res[0])
-        if img.res[1] > 0:
-            assert yy0[0] == pytest.approx(img.bounds.bottom)
-            assert yy0[-1] == pytest.approx(img.bounds.top - img.res[1])
-        else:
-            # Currently not covered by test image
-            assert yy0[0] == pytest.approx(img.bounds.top)
-            assert yy0[-1] == pytest.approx(img.bounds.bottom + img.res[1])
-
-        # With center argument
-        xx, yy = img.coords(grid=False, force_offset="center")
-        hx = img.res[0] / 2
-        hy = img.res[1] / 2
-        assert xx[0] == pytest.approx(img.bounds.left + hx)
-        assert xx[-1] == pytest.approx(img.bounds.right - hx)
-        if img.res[1] > 0:
-            assert yy[0] == pytest.approx(img.bounds.bottom + hy)
-            assert yy[-1] == pytest.approx(img.bounds.top - hy)
-        else:
-            # Currently not covered by test image
-            assert yy[-1] == pytest.approx(img.bounds.top + hy)
-            assert yy[0] == pytest.approx(img.bounds.bottom - hy)
-
-        # With grid argument (default argument, repeated here for clarity)
-        xxgrid, yygrid = img.coords(grid=True, force_offset="ll")
-        assert np.array_equal(xxgrid, np.repeat(xx0[np.newaxis, :], img.height, axis=0))
-        assert np.array_equal(yygrid, np.flipud(np.repeat(yy0[:, np.newaxis], img.width, axis=1)))
 
     @pytest.mark.parametrize("example", [landsat_b4_path, aster_dem_path, landsat_rgb_path])  # type: ignore
     def test_from_array(self, example: str) -> None:
